EXERCICIOS INICIAIS DE PYTHON

+++ 1. Tipos de dados /4. PAR E IMPAR.py
@@ -4,7 +4,7 @@ os.system("cls || clear")
 
 def main():
   # vamos pedir para o usuário informar um número
-  numero1 = int(input("Informe primeiro Número: "))
+  numero1 = float(input("Informe primeiro Número: "))
   
   # o número é par?
   if numero1 % 2 == 0:
